Move commands to app:* namespace.
